refactor(#78):paginationStore 생성 후 currentPage state 분리
