+++ src/app/(nonAuth)/community/page.tsx
@@ -4,7 +4,7 @@ import CommunityListSort from '@/components/communityList/CommunityListSort'
 import { ACTIVE_BUTTON_SHADOW } from '@/components/login/buttonCss'
 import { DOWN_ACTIVE_BUTTON } from '@/components/login/loginCss'
 import { getCommunityListInCommunity } from '@/query/community/communityQueryKey'
-import { usePaginationStore } from '@/shared/store/searchStore'
+import { usePaginationStore } from '@/shared/store/paginationStore'
 import Pagination from '@/util/Pagination '
 import { paging, resetPagination } from '@/util/util'
 import Link from 'next/link'
