+++ src/components/search/GenreMusicRecommendations.tsx
@@ -2,7 +2,7 @@
 
 import { GENRE_MUSIC_QUERY_KEY } from '@/query/genreMusic/queryKeys'
 import { getGenreMusicData } from '@/shared/main/api'
-import { usePaginationStore } from '@/shared/store/searchStore'
+import { usePaginationStore } from '@/shared/store/paginationStore'
 import Pagination from '@/util/Pagination '
 import { paging } from '@/util/util'
 import { useQuery } from '@tanstack/react-query'
