+++ src/components/search/SearchForm.tsx
@@ -3,10 +3,8 @@ import rectangle from '@/../public/images/Rectangle 156.svg'
 import arrowSearch from '@/../public/images/arrow_forward_ios.svg'
 import search from '@/../public/images/searchIcon (2).svg'
 import useInput from '@/hooks/useInput'
-import {
-  usePaginationStore,
-  useSearchedKeywordStore,
-} from '@/shared/store/searchStore'
+import { usePaginationStore } from '@/shared/store/paginationStore'
+import { useSearchedKeywordStore } from '@/shared/store/searchStore'
 import { resetPagination } from '@/util/util'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
