+++ src/shared/store/searchStore.ts
@@ -57,23 +57,3 @@ export const useModalMusicResultStore = create(
     },
   ),
 )
-
-type PaginationStore = {
-  currentPageData: {
-    currentPage: number
-  }
-  setCurrentPageData: (page: number) => void
-}
-
-const currentPageState = {
-  currentPageData: {
-    currentPage: 1,
-  },
-}
-
-export const usePaginationStore = create<PaginationStore>((set) => ({
-  ...currentPageState,
-  setCurrentPageData: (currentPage: number) => {
-    set({ currentPageData: { currentPage } })
-  },
-}))
